apps/python/requirements.txt
flask==3.0.3
flask-restful==0.3.10
requests==2.31.0
peewee==3.17.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
autopep8==2.1.0
flake8==7.0.0
flake8-pyproject==1.2.3
pytest==8.1.1
sympy==1.13.0
